requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pymongo==4.6.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic[email]==2.6.1
